Added Website Opener
Website Opener automatically opens all the website on your default browser. If someone uses multiple website and have to search each website or enter each URL address, then it is the program that can help. 
It stores sites in a database and opens each of them, database can be added and deleted as required.